+++ releasenotes.md
@@ -76,6 +76,7 @@
 - Debug info with recursive canonical type usage could cause segfault.
 - Missing check on optional left hand side for `s.x`.
 - Incorrect zero analysis on `foo["test"] = {}` #1360.
+- Bug converting untyped list #1360.
 
 ### Stdlib changes